api/views/reviewviewset.py
from django.shortcuts import get_object_or_404
from rest_framework import viewsets

from api.models import Title
from api.permission import IsAuthorOrReadOnly
from api.serializers.reviewserializer import ReviewSerializer


class ReviewViewSet(viewsets.ModelViewSet):
    permission_classes = [IsAuthorOrReadOnly, ]
    serializer_class = ReviewSerializer

    def get_queryset(self):
        title = get_object_or_404(Title, pk=self.kwargs.get('title_id'))
        return title.review.all()

    def perform_create(self, serializer):
        title = get_object_or_404(Title, pk=self.kwargs.get('title_id'))
        serializer.save(author=self.request.user, title=title)
